Esfera con cuda :)

+++ RaytracingEngine/interval.h
@@ -1,36 +1,37 @@
+#include "rtweekend.h"
 #ifndef INTERVAL_H
 #define INTERVAL_H
 
 class interval {
 public:
-    double min, max;
+    float min, max;
 
     interval() : min(+infinity), max(-infinity) {} // Default interval is empty
 
-    interval(double min, double max) : min(min), max(max) {}
+    interval(float min, float max) : min(min), max(max) {}
 
     interval(const interval& a, const interval& b) {
         // Create the interval tightly enclosing the two input intervals.
         min = a.min <= b.min ? a.min : b.min;
         max = a.max >= b.max ? a.max : b.max;
     }
-    double size() const {
+    float size() const {
         return max - min;
     }
 
-    bool contains(double x) const {
+    bool contains(float x) const {
         return min <= x && x <= max;
     }
 
-    bool surrounds(double x) const {
+    bool surrounds(float x) const {
         return min < x && x < max;
     }
-    double clamp(double x) const {
+    float clamp(float x) const {
         if (x < min) return min;
         if (x > max) return max;
         return x;
     }
-    interval expand(double delta) const {
+    interval expand(float delta) const {
         auto padding = delta/2;
         return interval(min - padding, max + padding);
     }
@@ -42,11 +43,11 @@ const interval interval::empty = interval(+infinity, -infinity);
 const interval interval::universe = interval(-infinity, +infinity);
 
 
-interval operator+(const interval& ival, double displacement) {
+interval operator+(const interval& ival, float displacement) {
     return interval(ival.min + displacement, ival.max + displacement);
 }
 
-interval operator+(double displacement, const interval& ival) {
+interval operator+(float displacement, const interval& ival) {
     return ival + displacement;
 }
 #endif
